Role refactoring
