+++ src/main/java/org/example/service/impl/RoleServiceImpl.java
@@ -9,7 +9,7 @@ import org.example.servlet.dto.RoleIncomingDto;
 import org.example.servlet.dto.RoleOutGoingDto;
 import org.example.servlet.dto.RoleUpdateDto;
 import org.example.servlet.mapper.RoleDtoMapper;
-import org.example.servlet.mapper.RoleDtoMapperImpl;
+import org.example.servlet.mapper.impl.RoleDtoMapperImpl;
 
 import java.util.List;
